% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam4r.r
\name{description}
\alias{description}
\title{Describe a Redatam object (database, entity or variable), as seen by libredatam}
\usage{
description(x, ...)
}
\arguments{
\item{obj}{Object to describe}
}
\description{
Describe a Redatam object (database, entity or variable), as seen by libredatam
}
