% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam4r.r
\name{read.redatam}
\alias{read.redatam}
\title{Read a Redatam database}
\usage{
read.redatam(dic_path)
}
\arguments{
\item{dic_path}{Path to the .dic file of the database}
}
\description{
Read a Redatam database
}
