% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam4r.r
\name{[.redatam.variable.chr}
\alias{[.redatam.variable.chr}
\title{Index a Redatam CHR variable. Triggers a RBF file read}
\usage{
\method{[}{redatam.variable.chr}(x, ...)
}
\arguments{
\item{x}{Object to index}

\item{i}{Indices (can be numeric, integer or boolean)}
}
\value{
Returns a character vector with the elements from the RBF file.
}
\description{
Index a Redatam CHR variable. Triggers a RBF file read
}
