% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam4r.r
\name{[.redatam.variable.int}
\alias{[.redatam.variable.int}
\title{Index a Redatam INT variable. Triggers a RBF file read}
\usage{
\method{[}{redatam.variable.int}(x, ...)
}
\arguments{
\item{x}{Object to index}

\item{i}{Indices (can be numeric, integer or boolean)}
}
\value{
Returns an integer vector (which can have factors) with the elements from the RBF file.
}
\description{
Index a Redatam INT variable. Triggers a RBF file read
}
