% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam4r.r
\name{[.redatam.variable.pck}
\alias{[.redatam.variable.pck}
\title{Index a Redatam PCK variable. Triggers a RBF file read. Note that the returned object size may be much greater than the size of the RBF file.}
\usage{
\method{[}{redatam.variable.pck}(x, ...)
}
\arguments{
\item{x}{Object to index}

\item{i}{Indices (can be numeric, integer or boolean)}
}
\value{
Returns an integer vector (which can have factors) with the elements from the RBF file.
}
\description{
Index a Redatam PCK variable. Triggers a RBF file read. Note that the returned object size may be much greater than the size of the RBF file.
}
